chore(product tours): add step order history to main app
